pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "unitrack"
version = "1.0.0"
description = "Universal Attendance Tracker for any university or institute"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "UniTrack", email = "unitrack@example.com"}
]
keywords = ["attendance", "university", "college", "tracker", "erp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Education",
]
dependencies = [
    "playwright>=1.40.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
unitrack = "unitrack.cli.main:cli"

[project.urls]
Homepage = "https://github.com/unitrack/unitrack"
Documentation = "https://github.com/unitrack/unitrack#readme"
Repository = "https://github.com/unitrack/unitrack"

[tool.setuptools.packages.find]
where = ["."]
include = ["unitrack*"]
